% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{calibrateHarmDensity}
\alias{calibrateHarmDensity}
\title{Calibrate a harm density function}
\usage{
calibrateHarmDensity(count, popn, p_cd, pcc, gamma_relation, lb, ub, threshold)
}
\arguments{
\item{count}{Total number of events among the given population}

\item{popn}{Total number of persons in the given population}

\item{p_cd}{Proportion of given population that are current drinkers}

\item{pcc}{daily per capita consumption in grams}

\item{gamma_relation}{relationship between mean and standard deviation of the
gamma distribution that governs alcohol consumption.  By Kehoe, use 1.171 for
male populations and 1.258 for female populations.}

\item{lb}{lower bound of consumption}

\item{ub}{upper bound of consumption}

\item{threshold}{lower bound of consumption at which condition occurs}
}
\description{
Calibrate a harm density function
}
