% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{makeConditionalProbability}
\alias{makeConditionalProbability}
\title{Factory for loglinear conditional probability functions}
\usage{
makeConditionalProbability(slope, lb)
}
\description{
Factory for loglinear conditional probability functions
}
