% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{makeCumulativeHarmDistribution}
\alias{makeCumulativeHarmDistribution}
\title{Factory for a harm distribution function}
\usage{
makeCumulativeHarmDistribution(mass, risk, lb, ub, target)
}
\description{
Factory for a harm distribution function
}
