% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{makeHarmDensity}
\alias{makeHarmDensity}
\title{Factory for a harm density function}
\usage{
makeHarmDensity(mass = mass, risk = risk, lb = lb, ub = ub,
  target = target)
}
\description{
Factory for a harm density function
}
