% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{makeHarmDistribution}
\alias{makeHarmDistribution}
\title{Factory for a harm distribution function}
\usage{
makeHarmDistribution(mass, risk, lb, ub, rtarget)
}
\description{
Factory for a harm distribution function
}
