% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{makeNormalizedGamma}
\alias{makeNormalizedGamma}
\title{Factory for a gamma density normalized over the given interval to the given
drinking population}
\usage{
makeNormalizedGamma(p_cd, pcc, gamma_relation, lb, ub)
}
\arguments{
\item{p_cd}{Proportion of given population that are current drinkers}

\item{pcc}{daily per capita consumption in grams}

\item{gamma_relation}{relationship between mean and standard deviation of the
gamma distribution that governs alcohol consumption.  By Kehoe, use 1.171 for
male populations and 1.258 for female populations.}

\item{lb}{lower bound of consumption}

\item{lb}{upper bound of consumption}
}
\description{
Factory for a gamma density normalized over the given interval to the given
drinking population
}
