% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{makeTarget}
\alias{makeTarget}
\title{Make a target incidence for loglinear slope calibration}
\usage{
makeTarget(count, popn, p_cd)
}
\arguments{
\item{count}{Total number of events among the given population}

\item{popn}{Total number of persons in the given population}
}
\description{
Make a target incidence for loglinear slope calibration
}
